增加叮当统计相关接口实现

+++ src/main/java/org/buaa/ly/MyCar/logic/VehicleLogic.java
@@ -13,7 +13,7 @@ public interface VehicleLogic {
 
     Vehicle find(String number);
 
-    long count(Integer sid, Integer viid, Collection<Integer> status, boolean exclude);
+    int count(Integer sid, Integer viid, Collection<Integer> status, boolean exclude);
 
     List<Vehicle> find(Integer sid, Integer viid, Collection<Integer> status, boolean exclude);
